dev: add unsplash API and add  env for magic login
