+++ apiserver/.env.example
@@ -62,3 +62,6 @@ ENABLE_SIGNUP="1"
 
 # Enable Email/Password Signup
 ENABLE_EMAIL_PASSWORD="1"
+
+# Enable Magic link Login
+ENABLE_MAGIC_LINK_LOGIN="0"
