+++ apiserver/plane/api/views/external.py
@@ -2,9 +2,10 @@
 import requests
 
 # Third party imports
+import openai
 from rest_framework.response import Response
 from rest_framework import status
-import openai
+from rest_framework.permissions import AllowAny
 from sentry_sdk import capture_exception
 
 # Django imports
@@ -15,6 +16,7 @@ from .base import BaseAPIView
 from plane.api.permissions import ProjectEntityPermission
 from plane.db.models import Workspace, Project
 from plane.api.serializers import ProjectLiteSerializer, WorkspaceLiteSerializer
+from plane.utils.integrations.github import get_release_notes
 
 
 class GPTIntegrationEndpoint(BaseAPIView):
@@ -67,6 +69,47 @@ class GPTIntegrationEndpoint(BaseAPIView):
                 {"error": "Workspace or Project Does not exist"},
                 status=status.HTTP_400_BAD_REQUEST,
             )
+        except Exception as e:
+            capture_exception(e)
+            return Response(
+                {"error": "Something went wrong please try again later"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+
+class ReleaseNotesEndpoint(BaseAPIView):
+    def get(self, request):
+        try:
+            release_notes = get_release_notes()
+            return Response(release_notes, status=status.HTTP_200_OK)
+        except Exception as e:
+            capture_exception(e)
+            return Response(
+                {"error": "Something went wrong please try again later"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+
+class UnsplashEndpoint(BaseAPIView):
+
+    def get(self, request):
+        try:
+            query = request.GET.get("query", False)
+            page = request.GET.get("page", 1)
+            per_page = request.GET.get("per_page", 20)
+
+            url = (
+                f"https://api.unsplash.com/search/photos/?client_id={settings.UNSPLASH_ACCESS_KEY}&query={query}&page=${page}&per_page={per_page}"
+                if query
+                else f"https://api.unsplash.com/photos/?client_id={settings.UNSPLASH_ACCESS_KEY}&page={page}&per_page={per_page}"
+            )
+
+            headers = {
+                "Content-Type": "application/json",
+            }
+
+            resp = requests.get(url=url, headers=headers)
+            return Response(resp.json(), status=status.HTTP_200_OK)
         except Exception as e:
             capture_exception(e)
             return Response(
